Use react-mounter on Meteor 1.3+
If 'require' is available, use it to gather the dependencies
needed to mount the React component that wraps our templates,
namely 'react' and 'react-mounter'.

This dependencies are not required in 'package.js' because of
the warnings in Meteor Guide[1].

In addition, 'kadira:react-layout' and 'react' atmosphere packages
are removed to avoid conflicts with NPM packages when 'flow-routing'
is used in a Meteor 1.3+ application.

Unfortunately this means that applications written for prior versions
of Meteor should now explicitly include them.

[1] https://guide.meteor.com/writing-atmosphere-packages.html#peer-npm-dependencies