Add check for special error from meteor/ddp-rate-limiter
Fixes [this issue](https://github.com/meteor-useraccounts/core/issues/629).
Without checking for this error, when rate-limiting kicks in it throws an error because `fieldId` is set to "timeToReset" (and this field does not exist). The whole accounts-form then is disabled and unresponsive, without informing the user about what's going on (you can only see the error message in the client-side browser console).